+++ bittytax/conv/parsers/hecoinfo.py
@@ -12,6 +12,10 @@ def parse_hecoinfo(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(int(row_dict['UnixTimestamp']))
 
+    if row_dict['Status'] != '':
+        # Failed txns should not have a Value_OUT
+        row_dict['Value_OUT(HT)'] = 0
+
     if Decimal(row_dict['Value_IN(HT)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
                                                  data_row.timestamp,
@@ -80,6 +84,22 @@ DataParser(DataParser.TYPE_EXPLORER,
            row_handler=parse_hecoinfo)
 
 DataParser(DataParser.TYPE_EXPLORER,
+           "HecoInfo (HECO Transactions)",
+           ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
+            'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
+            'Historical $Price/HT', 'Status', 'ErrCode', 'Method'],
+           worksheet_name="HecoInfo",
+           row_handler=parse_hecoinfo)
+
+DataParser(DataParser.TYPE_EXPLORER,
+           "HecoInfo (HECO Transactions)",
+           ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
+            'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
+            'Historical $Price/HT', 'Status', 'ErrCode', 'Method', 'PrivateNote'],
+           worksheet_name="HecoInfo",
+           row_handler=parse_hecoinfo)
+
+DataParser(DataParser.TYPE_EXPLORER,
            "HecoInfo (HECO Internal Transactions)",
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'ParentTxFrom', 'ParentTxTo',
             'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(HT)', 'Value_OUT(HT)',
@@ -94,6 +114,8 @@ DataParser(DataParser.TYPE_EXPLORER,
             None, 'Historical $Price/HT', 'Status', 'ErrCode', 'Type', 'PrivateNote'],
            worksheet_name="HecoInfo",
            row_handler=parse_hecoinfo_internal)
+
+
 
 # Same header as Etherscan
 #DataParser(DataParser.TYPE_EXPLORER,